feat:post 엔티티 폴더명 domain -> entity로 통일
