+++ src/main/java/com/example/moro/app/auth/handler/OAuth2SuccessHandler.java
@@ -50,3 +50,4 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         response.getWriter().write(objectMapper.writeValueAsString(apiResponse));
     }
 }
+
