+++ src/main/java/com/example/moro/app/post/entity/Comment.java
@@ -1,28 +1,23 @@
-package com.example.moro.app.post.domain;
+package com.example.moro.app.post.entity;
 
-import com.example.moro.app.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
-@Table(name = "like")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Like {
-
+public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="LikeId")
+    @Column(name = "cmmtId")
     private Long id;
+
+    @Column
+    private String content;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "postId")
     private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "userId")
-    private Member member;
 }
