+++ src/main/java/com/example/moro/app/post/entity/Like.java
@@ -1,24 +1,27 @@
-package com.example.moro.app.post.domain;
+package com.example.moro.app.post.entity;
 
+import com.example.moro.app.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
+@Table(name = "like")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment {
+public class Like {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "cmmtId")
+    @Column(name="LikeId")
     private Long id;
-
-    @Column
-    private String content;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "postId")
     private Post post;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "userId")
+    private Member member;
 }
