+++ src/main/java/com/example/moro/app/post/entity/PostColor.java
@@ -1,12 +1,10 @@
-package com.example.moro.app.post.domain;
+package com.example.moro.app.post.entity;
 
 import com.example.moro.app.colormap.entity.ColorMap;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.awt.*;
 
 @Entity
 @Getter
