+++ src/main/java/com/example/moro/app/post/repository/PostRepository.java
@@ -1,0 +1,10 @@
+package com.example.moro.app.post.repository;
+
+import com.example.moro.app.post.entity.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post,Long> {
+
+}
